Rename file web-snippet.vue to WebSnippet.vue
