+++ components/slices/SliceZone.vue
@@ -16,7 +16,7 @@
 
 <script>
 import MultipleColumn from '@/components/slices/MultipleColumn'
-import WebSnippet from '~/components/slices/web-snippet'
+import WebSnippet from '@/components/slices/WebSnippet'
 
 export default {
   name: 'SliceZone',
